Allow installing holmes locally with pip
