+++ pyproject.toml
@@ -4,6 +4,10 @@ version = "0.1.0"
 description = ""
 authors = ["Natan Yellin <natan@robusta.dev>"]
 readme = "README.md"
+packages = [{ include = "holmes" }]
+
+[tool.poetry.scripts]
+holmes = "holmes.main:run"
 
 [tool.poetry.dependencies]
 python = "^3.9"
